requirements.txt
Scrapy==0.24.5
scrapylib==1.5.0
python-dateutil==2.4.2
pickledb==0.3